requirements/development.txt
-r base.txt

psycopg2-binary==2.8.6

pytest==6.2.4
pytest-django==4.3.0
factory-boy==3.2.0
django-debug-toolbar==3.2.1

black==20.8b1
pre-commit==2.12.0